+++ src/main/java/com/jangburich/domain/entity/FavoriteTeam.java
@@ -1,4 +1,4 @@
-package com.jangburich.domain.team.domain;
+package com.jangburich.domain.entity;
 
 import com.jangburich.domain.common.BaseEntity;
 import com.jangburich.domain.user.domain.User;
@@ -11,14 +11,13 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class UserTeam extends BaseEntity {
+public class FavoriteTeam extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -32,16 +31,4 @@ public class UserTeam extends BaseEntity {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "team_id")
     private Team team;
-
-    public UserTeam(User user, Team team) {
-        this.user = user;
-        this.team = team;
-    }
-
-    public static UserTeam of(User user, Team team) {
-        if (user == null || team == null) {
-            throw new IllegalArgumentException("유저와 팀은 null이 될 수 없습니다.");
-        }
-        return new UserTeam(user, team);
-    }
 }
