+++ src/main/java/com/jangburich/domain/entity/Orders.java
@@ -5,7 +5,6 @@ import java.math.BigDecimal;
 import org.hibernate.annotations.Comment;
 
 import com.jangburich.domain.common.BaseEntity;
-import com.jangburich.domain.team.domain.Team;
 import com.jangburich.domain.user.domain.User;
 
 import jakarta.persistence.Column;
