+++ src/main/java/com/jangburich/domain/point/domain/PointTransaction.java
@@ -2,7 +2,7 @@ package com.jangburich.domain.point.domain;
 
 import com.jangburich.domain.common.BaseEntity;
 import com.jangburich.domain.entity.Store;
-import com.jangburich.domain.team.domain.Team;
+import com.jangburich.domain.entity.Team;
 import com.jangburich.domain.user.domain.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
