+++ src/main/java/com/jangburich/domain/repository/TeamRepository.java
@@ -1,11 +1,11 @@
-package com.jangburich.domain.team.domain.repository;
+package com.jangburich.domain.repository;
 
 import com.jangburich.domain.common.Status;
-import com.jangburich.domain.team.domain.Team;
+import com.jangburich.domain.entity.Team;
+import com.jangburich.domain.repository.queryDsl.TeamQueryDslRepository;
 import com.jangburich.domain.user.domain.User;
 import java.util.List;
 import java.util.Optional;
-import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
